+++ 3D/functions_AECM.hpp
@@ -423,7 +423,7 @@ void AECM_optim(Matrix_eig_row &W_init, Matrix3d_eig &Psi_inv, Vector_eig &beta,
 		}
 	}
 	if(verbose)
-		Debug0("Number of possible background voxels: " << black_list.sum());
+		Debug0("Number of possible background voxels: " << (unsigned long int)(black_list.sum()));
 	
 	
 	
@@ -436,7 +436,7 @@ void AECM_optim(Matrix_eig_row &W_init, Matrix3d_eig &Psi_inv, Vector_eig &beta,
 		}
 	}
 	if(verbose)
-		Debug0("Number of possible checkerboard white ones: " << checkerboard_white.sum());
+		Debug0("Number of possible checkerboard white ones: " << (unsigned long int)(checkerboard_white.sum()));
